docs: update docs
This updates the docs and performs some minor refactoring.
