+++ webauthn/registration.go
@@ -8,12 +8,7 @@ import (
 	"time"
 
 	"github.com/go-webauthn/webauthn/protocol"
-	"github.com/go-webauthn/webauthn/protocol/webauthncose"
 )
-
-// BEGIN REGISTRATION
-// These objects help us create the CredentialCreationOptions
-// that will be passed to the authenticator via the user client.
 
 // RegistrationOption describes a function which modifies the registration
 // [*protocol.PublicKeyCredentialCreationOptions] values.
@@ -113,105 +108,6 @@ func (webauthn *WebAuthn) BeginMediatedRegistration(user User, mediation protoco
 	return creation, session, nil
 }
 
-// WithCredentialParameters adjusts the credential parameters in the registration options.
-func WithCredentialParameters(credentialParams []protocol.CredentialParameter) RegistrationOption {
-	return func(cco *protocol.PublicKeyCredentialCreationOptions) {
-		cco.Parameters = credentialParams
-	}
-}
-
-// WithExclusions adjusts the non-default parameters regarding credentials to exclude from registration.
-func WithExclusions(excludeList []protocol.CredentialDescriptor) RegistrationOption {
-	return func(cco *protocol.PublicKeyCredentialCreationOptions) {
-		cco.CredentialExcludeList = excludeList
-	}
-}
-
-// WithAuthenticatorSelection adjusts the non-default parameters regarding the authenticator to select during
-// registration.
-func WithAuthenticatorSelection(authenticatorSelection protocol.AuthenticatorSelection) RegistrationOption {
-	return func(cco *protocol.PublicKeyCredentialCreationOptions) {
-		cco.AuthenticatorSelection = authenticatorSelection
-	}
-}
-
-// WithResidentKeyRequirement sets both the resident key and require resident key protocol options.
-func WithResidentKeyRequirement(requirement protocol.ResidentKeyRequirement) RegistrationOption {
-	return func(cco *protocol.PublicKeyCredentialCreationOptions) {
-		cco.AuthenticatorSelection.ResidentKey = requirement
-
-		switch requirement {
-		case protocol.ResidentKeyRequirementRequired:
-			cco.AuthenticatorSelection.RequireResidentKey = protocol.ResidentKeyRequired()
-		default:
-			cco.AuthenticatorSelection.RequireResidentKey = protocol.ResidentKeyNotRequired()
-		}
-	}
-}
-
-// WithPublicKeyCredentialHints adjusts the non-default hints for credential types to select during registration.
-//
-// WebAuthn Level 3.
-func WithPublicKeyCredentialHints(hints []protocol.PublicKeyCredentialHints) RegistrationOption {
-	return func(cco *protocol.PublicKeyCredentialCreationOptions) {
-		cco.Hints = hints
-	}
-}
-
-// WithConveyancePreference adjusts the non-default parameters regarding whether the authenticator should attest to the
-// credential.
-func WithConveyancePreference(preference protocol.ConveyancePreference) RegistrationOption {
-	return func(cco *protocol.PublicKeyCredentialCreationOptions) {
-		cco.Attestation = preference
-	}
-}
-
-// WithAttestationFormats adjusts the non-default formats for credential types to select during registration.
-//
-// WebAuthn Level 3.
-func WithAttestationFormats(formats []protocol.AttestationFormat) RegistrationOption {
-	return func(cco *protocol.PublicKeyCredentialCreationOptions) {
-		cco.AttestationFormats = formats
-	}
-}
-
-// WithExtensions adjusts the extension parameter in the registration options.
-func WithExtensions(extension protocol.AuthenticationExtensions) RegistrationOption {
-	return func(cco *protocol.PublicKeyCredentialCreationOptions) {
-		cco.Extensions = extension
-	}
-}
-
-// WithAppIdExcludeExtension automatically includes the specified appid if the CredentialExcludeList contains a credential
-// with the type `fido-u2f`.
-func WithAppIdExcludeExtension(appid string) RegistrationOption {
-	return func(cco *protocol.PublicKeyCredentialCreationOptions) {
-		for _, credential := range cco.CredentialExcludeList {
-			if credential.AttestationType == protocol.CredentialTypeFIDOU2F {
-				if cco.Extensions == nil {
-					cco.Extensions = map[string]any{}
-				}
-
-				cco.Extensions[protocol.ExtensionAppIDExclude] = appid
-			}
-		}
-	}
-}
-
-// WithRegistrationRelyingPartyID sets the relying party id for the registration.
-func WithRegistrationRelyingPartyID(id string) RegistrationOption {
-	return func(cco *protocol.PublicKeyCredentialCreationOptions) {
-		cco.RelyingParty.ID = id
-	}
-}
-
-// WithRegistrationRelyingPartyName sets the relying party name for the registration.
-func WithRegistrationRelyingPartyName(name string) RegistrationOption {
-	return func(cco *protocol.PublicKeyCredentialCreationOptions) {
-		cco.RelyingParty.Name = name
-	}
-}
-
 // FinishRegistration takes the response from the authenticator and client and verify the credential against the user's
 // credentials and session data.
 //
@@ -252,115 +148,4 @@ func (webauthn *WebAuthn) CreateCredential(user User, session SessionData, parse
 	}
 
 	return NewCredential(clientDataHash, parsedResponse)
-}
-
-// CredentialParametersDefault is the default [protocol.CredentialParameter] list.
-func CredentialParametersDefault() []protocol.CredentialParameter {
-	return []protocol.CredentialParameter{
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgES256,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgES384,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgES512,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgRS256,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgRS384,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgRS512,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgPS256,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgPS384,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgPS512,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgEdDSA,
-		},
-	}
-}
-
-// CredentialParametersRecommendedL3 is explicitly the Level 3 recommended [protocol.CredentialParameter] list.
-func CredentialParametersRecommendedL3() []protocol.CredentialParameter {
-	return []protocol.CredentialParameter{
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgEdDSA,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgES256,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgRS256,
-		},
-	}
-}
-
-// CredentialParametersExtendedL3 is the Level 3 recommended [protocol.CredentialParameter] list with all of the other
-// parameters supported by the library.
-func CredentialParametersExtendedL3() []protocol.CredentialParameter {
-	return []protocol.CredentialParameter{
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgEdDSA,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgES256,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgES384,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgES512,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgRS256,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgRS384,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgRS512,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgPS256,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgPS384,
-		},
-		{
-			Type:      protocol.PublicKeyCredentialType,
-			Algorithm: webauthncose.AlgPS512,
-		},
-	}
 }
